Changes made so code compiles and runs without crashing
A few dummy functions (GameState.java: tempDeck and clearTemp) and two commented out function calls:

game.run() in Test.java
forwardModel.setup(gameState) in PandemicGame.java

+++ src/main/java/content/PropertyIntArray.java
@@ -26,7 +26,7 @@ public class PropertyIntArray extends Property
         this.values = new int[values.size()];
         for(int i =0; i< values.size(); ++i)
         {
-            this.values[i] = (int) values.get(i);
+            this.values[i] = ((Long) values.get(i)).intValue();
         }
 
     }
